7_Image Captioning Model for Football Highlights (2025)/requirements.txt
# Core ML/DL libraries
tensorflow>=2.10.0
keras>=2.10.0
torch>=1.12.0
torchvision>=0.13.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0
opencv-python>=4.7.0
Pillow>=9.3.0
scikit-learn>=1.1.0

# Natural Language Processing
nltk>=3.7
rouge-score>=0.1.2
transformers>=4.21.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# Web application
streamlit>=1.15.0
flask>=2.2.0

# Data collection
requests>=2.28.0
flickrapi>=2.4.0
kaggle>=1.5.12

# Utilities
tqdm>=4.64.0
python-dotenv>=0.19.0
jupyter>=1.0.0
notebook>=6.4.0

# Image processing
scikit-image>=0.19.0
albumentations>=1.3.0

# Evaluation metrics
evaluate>=0.4.0
jiwer>=3.0.0

# Development tools
black>=22.0.0
flake8>=4.0.0
pytest>=7.1.0

# Optional: GPU support (uncomment if you have CUDA-compatible GPU)
# tensorflow-gpu>=2.10.0